requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-allauth==64.0.0
django-crispy-forms==2.3
django-formtools==2.5.1
django-widget-tweaks==1.5.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
holidays==0.57
idna==3.7
joblib==1.4.2
numpy
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
stripe==10.10.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
whitenoise
gunicorn
web3==6.11.1
eth-account==0.8.0
eth-typing==3.5.2
eth-utils==2.3.1
hexbytes==0.3.1
